# Get rid of boost::math::constants::* and M_PI in favor of std::numbers