--- a/include/util/cheap_ruler.hpp
+++ b/include/util/cheap_ruler.hpp
@@ -4,9 +4,9 @@
 #include <cmath>
 #include <cstdint>
 #include <limits>
+#include <numbers>
 #include <tuple>
 #include <utility>
-#include <numbers>
 
 namespace mapbox
 {