--- a/include/util/trigonometry_table.hpp
+++ b/include/util/trigonometry_table.hpp
@@ -6,7 +6,7 @@
 
 #include <limits>
 
-#include <boost/math/constants/constants.hpp>
+#include <numbers>
 
 namespace osrm::util
 {
@@ -359,22 +359,22 @@ constexpr unsigned short atan_table[4096] = {
 #ifdef _MSC_VER // TODO: remove as soon as boost allows C++14 features with Visual Studio
 const constexpr double SCALING_FACTOR = 4. / M_PI * 0xFFFF;
 #else
-const constexpr double SCALING_FACTOR = 4. / boost::math::constants::pi<double>() * 0xFFFF;
+const constexpr double SCALING_FACTOR = 4. / std::numbers::pi * 0xFFFF;
 #endif
 
 inline double atan2_lookup(double y, double x)
 {
-    using namespace boost::math::constants;
+    static constexpr auto half_pi = std::numbers::pi * 0.5;
 
     if (std::abs(x) < std::numeric_limits<double>::epsilon())
     {
         if (y >= 0.)
         {
-            return half_pi<double>();
+            return half_pi;
         }
         else
         {
-            return -half_pi<double>();
+            return -half_pi;
         }
     }
 
@@ -405,25 +405,25 @@ inline double atan2_lookup(double y, double x)
     case 0:
         break;
     case 1:
-        angle = pi<double>() - angle;
+        angle = std::numbers::pi - angle;
         break;
     case 2:
         angle = -angle;
         break;
     case 3:
-        angle = -pi<double>() + angle;
+        angle = -std::numbers::pi + angle;
         break;
     case 4:
-        angle = half_pi<double>() - angle;
+        angle = half_pi - angle;
         break;
     case 5:
-        angle = half_pi<double>() + angle;
+        angle = half_pi + angle;
         break;
     case 6:
-        angle = -half_pi<double>() + angle;
+        angle = -half_pi + angle;
         break;
     case 7:
-        angle = -half_pi<double>() - angle;
+        angle = -half_pi - angle;
         break;
     }
     return angle;