--- a/include/util/trigonometry_table.hpp
+++ b/include/util/trigonometry_table.hpp
@@ -359,7 +359,7 @@ constexpr unsigned short atan_table[4096] = {
 #ifdef _MSC_VER // TODO: remove as soon as boost allows C++14 features with Visual Studio
 const constexpr double SCALING_FACTOR = 4. / M_PI * 0xFFFF;
 #else
-const constexpr double SCALING_FACTOR = 4. / std::numbers::pi * 0xFFFF;
+const constexpr double SCALING_FACTOR = 4. * std::numbers::inv_pi * 0xFFFF;
 #endif
 
 inline double atan2_lookup(double y, double x)