Let's prefer things from stdlib...

<!-- BENCHMARK_RESULTS_START -->
## Benchmark Results
| Benchmark | Base | PR |
|-----------|------|----|
| alias | aliased u32: 1105.94<br/>plain u32: 1098.95<br/>aliased double: 959.533<br/>plain double: 964.602 | aliased u32: 1097<br/>plain u32: 1101.59<br/>aliased double: 961.788<br/>plain double: 960.653 |
| json-render | String: 6.56501ms<br/>Stringstream: 9.29638ms<br/>Vector: 6.89917ms | String: 6.60594ms<br/>Stringstream: 9.31658ms<br/>Vector: 6.95441ms |
| match_ch | Default radius: <br/>4.44353ms/req at 82 coordinate<br/>0.0541894ms/coordinate<br/>Radius 5m: <br/>4.41448ms/req at 82 coordinate<br/>0.0538351ms/coordinate<br/>Radius 10m: <br/>15.1272ms/req at 82 coordinate<br/>0.184478ms/coordinate<br/>Radius 15m: <br/>37.1259ms/req at 82 coordinate<br/>0.452755ms/coordinate<br/>Radius 30m: <br/>315.501ms/req at 82 coordinate<br/>3.84757ms/coordinate | Default radius: <br/>4.74375ms/req at 82 coordinate<br/>0.0578506ms/coordinate<br/>Radius 5m: <br/>4.75225ms/req at 82 coordinate<br/>0.0579542ms/coordinate<br/>Radius 10m: <br/>16.3889ms/req at 82 coordinate<br/>0.199864ms/coordinate<br/>Radius 15m: <br/>40.2069ms/req at 82 coordinate<br/>0.490328ms/coordinate<br/>Radius 30m: <br/>343.263ms/req at 82 coordinate<br/>4.18613ms/coordinate |
| match_mld | Default radius: <br/>2.85775ms/req at 82 coordinate<br/>0.0348506ms/coordinate<br/>Radius 5m: <br/>2.99267ms/req at 82 coordinate<br/>0.036496ms/coordinate<br/>Radius 10m: <br/>10.532ms/req at 82 coordinate<br/>0.128439ms/coordinate<br/>Radius 15m: <br/>26.2624ms/req at 82 coordinate<br/>0.320274ms/coordinate<br/>Radius 30m: <br/>308.441ms/req at 82 coordinate<br/>3.76147ms/coordinate | Default radius: <br/>2.92088ms/req at 82 coordinate<br/>0.0356205ms/coordinate<br/>Radius 5m: <br/>2.91631ms/req at 82 coordinate<br/>0.0355647ms/coordinate<br/>Radius 10m: <br/>11.0469ms/req at 82 coordinate<br/>0.134718ms/coordinate<br/>Radius 15m: <br/>28.7166ms/req at 82 coordinate<br/>0.350202ms/coordinate<br/>Radius 30m: <br/>340.643ms/req at 82 coordinate<br/>4.15419ms/coordinate |
| packedvector | random write:<br/>std::vector 12113.6 ms<br/>util::packed_vector 82060.9 ms<br/>slowdown: 6.77429<br/>random read:<br/>std::vector 12158.6 ms<br/>util::packed_vector 33880.9 ms<br/>slowdown: 2.78659 | random write:<br/>std::vector 9872.18 ms<br/>util::packed_vector 81795.8 ms<br/>slowdown: 8.28549<br/>random read:<br/>std::vector 8576.45 ms<br/>util::packed_vector 33545 ms<br/>slowdown: 3.91129 |
| route_ch | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>506.633ms<br/>0.506633ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>347.475ms<br/>0.347475ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>625.175ms<br/>0.625175ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>152.014ms<br/>0.152014ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>97.3768ms<br/>0.0973768ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>132.47ms<br/>0.13247ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>151.173ms<br/>0.151173ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>97.3354ms<br/>0.0973354ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>132.317ms<br/>0.132317ms/req | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>510.971ms<br/>0.510971ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>351.5ms<br/>0.3515ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>630.938ms<br/>0.630938ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>151.031ms<br/>0.151031ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>97.1235ms<br/>0.0971235ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>132.182ms<br/>0.132182ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>150.397ms<br/>0.150397ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>97.1684ms<br/>0.0971684ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>131.916ms<br/>0.131916ms/req |
| route_mld | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>642.159ms<br/>0.642159ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>438.686ms<br/>0.438686ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>808.316ms<br/>0.808316ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>286.203ms<br/>0.286203ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>163.405ms<br/>0.163405ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>299.09ms<br/>0.29909ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>277.855ms<br/>0.277855ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>162.643ms<br/>0.162643ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>294.36ms<br/>0.29436ms/req | 1000 routes, 3 coordinates, no alternatives, overview=full, steps=true<br/>655.219ms<br/>0.655219ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=full, steps=true<br/>446.499ms<br/>0.446499ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=full, steps=true<br/>834.273ms<br/>0.834273ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false<br/>298.316ms<br/>0.298316ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false<br/>164.838ms<br/>0.164838ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false<br/>316.454ms<br/>0.316454ms/req<br/>1000 routes, 3 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>297.283ms<br/>0.297283ms/req<br/>1000 routes, 2 coordinates, no alternatives, overview=false, steps=false, radius=750<br/>164.39ms<br/>0.16439ms/req<br/>1000 routes, 2 coordinates, 3 alternatives, overview=false, steps=false, radius=750<br/>310.592ms<br/>0.310592ms/req |
| rtree | 1 result:<br/>207.759ms ->  0.0207759 ms/query<br/>10 results:<br/>242.618ms ->  0.0242618 ms/query | 1 result:<br/>209.939ms ->  0.0209939 ms/query<br/>10 results:<br/>243.93ms ->  0.024393 ms/query |
<!-- BENCHMARK_RESULTS_END -->