@SiarheiFedartsou 2024-07-08
@db4 yes, we migrated to C++20 recently